+++ intake_esm/aggregate.py
@@ -85,6 +85,13 @@ def merge(dsets):
     return dsm
 
 
+def _restore_non_dim_coords(ds):
+    """restore non_dim_coords to variables"""
+    non_dim_coords_reset = set(ds.coords) - set(ds.dims)
+    ds = ds.reset_coords(non_dim_coords_reset)
+    return ds
+
+
 def concat_time_levels(dsets, time_coord_name_default):
     """
     Concatenate datasets across "time" levels, taking time invariant variables
@@ -125,6 +132,8 @@ def concat_time_levels(dsets, time_coord_name_default):
     objs_to_concat = [first] + rest
 
     ds = xr.concat(objs_to_concat, dim=time_coord_name, coords='minimal')
+
+    ds = _restore_non_dim_coords(ds)
 
     new_history = f"\n{datetime.now()} xarray.concat(<ALL_TIMESTEPS>, dim='{time_coord_name}', coords='minimal')"
     if 'history' in attrs:
@@ -167,9 +176,7 @@ def concat_ensembles(
     ensemble_dim = xr.DataArray(member_ids, dims=ensemble_dim_name, name=ensemble_dim_name)
     ds = xr.concat(objs_to_concat, dim=ensemble_dim, coords='minimal')
 
-    # restore non_dim_coords to variables
-    non_dim_coords_reset = set(ds.coords) - set(ds.dims)
-    ds = ds.reset_coords(non_dim_coords_reset)
+    ds = _restore_non_dim_coords(ds)
 
     new_history = (
         f"\n{datetime.now()} xarray.concat(<ALL_MEMBERS>, dim='member_id', coords='minimal')"